Consolidating signature store and watcher db
