+++ mobilecoind/src/service.rs
@@ -825,11 +825,6 @@ impl<T: BlockchainConnection + UserTxConnection + 'static> ServiceApi<T> {
                 .map_err(|err| {
                     rpc_internal_error("watcher_db.get_block_signatures", err, &self.logger)
                 })?;
-            println!(
-                "\x1b[1;33m Got signatures len: {:?} for block index {:?}\x1b[0m",
-                signatures.len(),
-                request.block
-            );
             for signature in signatures.iter() {
                 response
                     .mut_signatures()
