+++ watcher/src/error.rs
@@ -2,43 +2,6 @@
 
 use failure::Fail;
 
-/// SignatureStore Errors
-#[derive(Debug, Eq, PartialEq, Copy, Clone, Fail)]
-pub enum SignatureStoreError {
-    #[fail(display = "NotFound")]
-    NotFound,
-
-    #[fail(display = "Serialization")]
-    Serialization,
-
-    #[fail(display = "Deserialization")]
-    Deserialization,
-
-    #[fail(display = "LmdbError: {}", _0)]
-    LmdbError(lmdb::Error),
-}
-
-impl From<prost::DecodeError> for SignatureStoreError {
-    fn from(_src: prost::DecodeError) -> Self {
-        SignatureStoreError::Deserialization
-    }
-}
-
-impl From<prost::EncodeError> for SignatureStoreError {
-    fn from(_src: prost::EncodeError) -> Self {
-        SignatureStoreError::Serialization
-    }
-}
-
-impl From<lmdb::Error> for SignatureStoreError {
-    fn from(src: lmdb::Error) -> Self {
-        match src {
-            lmdb::Error::NotFound => SignatureStoreError::NotFound,
-            _ => SignatureStoreError::LmdbError(src),
-        }
-    }
-}
-
 /// WatcherDB Errors
 #[derive(Debug, Eq, PartialEq, Copy, Clone, Fail)]
 pub enum WatcherDBError {
@@ -56,9 +19,6 @@ pub enum WatcherDBError {
 
     #[fail(display = "LmdbError: {}", _0)]
     LmdbError(lmdb::Error),
-
-    #[fail(display = "SignatureStore: {}", _0)]
-    SignatureStore(SignatureStoreError),
 }
 
 impl From<lmdb::Error> for WatcherDBError {
@@ -67,14 +27,14 @@ impl From<lmdb::Error> for WatcherDBError {
     }
 }
 
-impl From<SignatureStoreError> for WatcherDBError {
-    fn from(src: SignatureStoreError) -> Self {
-        WatcherDBError::SignatureStore(src)
-    }
-}
-
 impl From<prost::DecodeError> for WatcherDBError {
     fn from(_src: prost::DecodeError) -> Self {
         WatcherDBError::Deserialization
+    }
+}
+
+impl From<prost::EncodeError> for WatcherDBError {
+    fn from(_src: prost::EncodeError) -> Self {
+        WatcherDBError::Serialization
     }
 }
